requirements.txt
requests>=2.31.0
orjson>=3.9.0
Pillow>=10.0.0
numpy>=1.24.0
roman>=4.0
aiohttp>=3.8.0
bs4>=0.0.2
lxml>=4.6.3
jsondiff>=2.0.0